requirements.txt
numpy==1.23.4
pillow==9.2.0
